Create component and style it
The component returns two divs at the moment. The first one is used to
define it as flex and indicate it to strecth the remaining height. The
second one is defined with a grid display to accomodate the upcoming
cards of the game.

On the other hand, in the index.css file the root div in the html file
was targeted to indicate it to cover the whole screen's height.