Render cards in Gameboard components
At the moment 5 images have been imported for the game. These images
were arranged in objects, each one representing an element of an array
that is passed in to the card component, which is inside a map function.

The objects for the images contain the name of the image, and the
variable containing the link.