Shuffle cards on the screen
Every time a card is clicked the shuffleCards function is called. At the
outset, a new variable was declared containing the array of images and
texts; however, this was not a copy but a reference. To solve this, the
following syntax was followed to create a copy of the images state:

    let imageArrays = [...images]

On the other hand, previously two arrays states were created to store
the image path and its corresponding text, but this was changed for a
single array containing objects of the image and its text.