Create logic to evaluate cards
The array containing the card's information was added a new key to each
object, which is the id. Each image has an ID associated in order to
catch that value and evaluate whether that number has already been
stored in an upcoming state to define in the App component.

The function changeScore was defined in the App component and passed in
to the Gameboard component, so it is called within the shuffleCards
function. On the other hand, the logic for shuffling the cards is
executed after 0.1 ms, otherwise the changeScore function receives
the ID of the next card.