Add logic to increase Score or reset it
Once a card is clicked, its ID is evaluated to see if it is included
within the new state created in the App component, which is clickedCards
containing an empty array by default. If the ID is not included, then
the clickedCards are updated to contain that value and the current Score
increments by one.

Conversely, when the ID is included in clickedCards, then the value of
the current score is evaluated to see if the best Score has been beaten
to update its content. After that, the current Score is cleared as well
as the clickedCards state.